Refactor: moved utils.py to style_bert_vits2/models/
