+++ style_gen.py
@@ -6,8 +6,8 @@ import numpy as np
 import torch
 from tqdm import tqdm
 
-import utils
 from style_bert_vits2.logging import logger
+from style_bert_vits2.models import utils
 from style_bert_vits2.utils.stdout_wrapper import SAFE_STDOUT
 from config import config
